pyproject.toml
[project]
name = "vn-aio-atlas"
version = "0.1.0"
description = "Empirical study of Google AI Overview behavior in Vietnamese commercial search."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Hoang Duc Viet" }]
requires-python = ">=3.11"

dependencies = [
    "polars>=1.20.0",
    "duckdb>=1.2.0",
    "pyarrow>=18.0.0",
    "psycopg[binary,pool]>=3.2.0",
    "python-dotenv>=1.0.1",
    "matplotlib>=3.9.0",
    "altair>=5.5.0",
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]
